% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock.R
\name{get_stock}
\alias{get_stock}
\alias{magic_eightball}
\title{Get the jumping rivers stock price}
\usage{
get_stock(current = NULL)

magic_eightball(stock)
}
\arguments{
\item{current}{A data}

\item{stock}{current stock price}
}
\description{
Get the jumping rivers stock price
}
